Rename `render` to `show document`
