+++ lib/inventoryware.rb
@@ -85,15 +85,6 @@ module Inventoryware
     action(c, Commands::Parse)
   end
 
-  command :render do |c|
-    cli_syntax(c, 'TEMPLATE NODE(S)')
-    c.description = "Render nodes' data as an eRuby template"
-    c.option '-l', '--location LOCATION',
-      "Output the rendered template to the specified location."
-    c = add_node_options(c)
-    action(c, Commands::Render)
-  end
-
   command :modify do |c|
     cli_syntax(c)
     c.description = 'Change mutable node data'
@@ -144,6 +135,16 @@ module Inventoryware
     c.description = "View the .yaml for a node"
     c.hidden = true
     action(c, Commands::Shows::Data)
+  end
+
+  command :'show document' do |c|
+    cli_syntax(c, 'TEMPLATE NODE(S)')
+    c.description = "Create a document using nodes' data and an eRuby template"
+    c.option '-l', '--location LOCATION',
+      "Output the rendered template to the specified location."
+    c = add_node_options(c)
+    c.hidden = true
+    action(c, Commands::Shows::Document)
   end
 
   command :delete do |c|
